Conflicts:
	com.gongpingjia.carplay/com.gongpingjia.carplay.common/src/main/java/com/gongpingjia/carplay/common/util/ToolsUtils.java

+++ com.gongpingjia.carplay/com.gongpingjia.carplay.service/src/main/java/com/gongpingjia/carplay/po/Car.java
@@ -11,7 +11,7 @@ public class Car {
 
     private String model;
 
-    private Long price;
+    private Double price;
 
     private String color;
 
@@ -59,11 +59,11 @@ public class Car {
         this.model = model;
     }
 
-    public Long getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
